==> notebooks/data_wrangling/01_wiki_notable_people_wrangling_1.ipynb <==
# Analyzing Wikipedia - Gender-Bias?

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px

#change display settings to show all dataframe columns as well as complete cell values (no truncating values)
#https://stackoverflow.com/questions/47022070/display-all-dataframe-columns-in-a-jupyter-python-notebook
pd.set_option('display.max_columns', None)

#https://stackoverflow.com/a/51540918
#pd.set_option('display.max_rows', 500)
#pd.set_option('display.width', 1000)
pd.set_option('display.expand_frame_repr', False)
pd.set_option('max_colwidth', None)

## Data Source

- **Project**: [A cross-verified database of notable people, 3500BC-2018AD](https://www.nature.com/articles/s41597-022-01369-4)
    - Paper: [Shareable Link](https://rdcu.be/c6YvW)
- **Download**: [A Brief History of Human Time - Cross-verified Dataset](https://data.sciencespo.fr/dataset.xhtml?persistentId=doi:10.21410/7E4/RDAG3O)

Includes data until 2018.

## Read Data

#!ls ../data/notable_people_cross-verified/

#read file using utf-8 encoding to accommodate "ä", "ö", ... , "ß" etc.

notable = pd.read_csv("../data/notable_people_cross-verified/cross-verified-database.csv", 
                      encoding='utf-8', index_col=0)

#result: all 2.29 million individuals imported

#checking no. of records, columns and data types

notable.info()

notable.sample(5)

notable["bigperiod_birth_graph_b"].unique()

notable["bigperiod_death_graph_b"].unique()

notable["bigperiod_birth"].unique()

notable["bigperiod_death"].unique()

notable["approx_birth"].unique()

notable["approx_death"].unique()

notable["pantheon_1"].unique()

##### Check that special characters have been imported correctly

notable.loc[notable["level3_all_occ"] == "D:_economist_P:_pädagoge_reformer_ökonom_German"]

notable[["name"]].loc[notable["name"] == "Walter_Schultheiß"]

notable[["level3_all_occ"]].loc[["Q10068"]]

## Check for NULL values

#checking no. of NULL values per column

notable.isnull().sum()

#notable.loc[notable["gender"].isnull() == True]

#notable.loc[notable["gender"].isna() == True]

## Drop Columns not necessary for planned Analysis

#notable["approx_birth"].value_counts()

notable.drop(columns=["updated_death_date",
                      "approx_birth", 
                      "approx_death", 
                      "birth_min", 
                      "birth_max", 
                      "death_min", 
                      "death_max", 
                      "birth_estimation", 
                      "death_estimation",
                      "bigperiod_birth",
                      "bigperiod_death",
                      "level2_second_occ",
                      "freq_main_occ",
                      "freq_second_occ",
                      "all_geography_groups",
                      "string_citizenship_raw_d",
                      "citizenship_2_b",
                      "list_areas_of_rattach",
                      "area1_of_rattachment",
                      "area2_of_rattachment",
                      "pantheon_1",
                      "list_wikipedia_editions"
                     ], inplace=True)

## Check Gender Data

#check which unique gender values exist in data

notable["gender"].unique()

#check no. of unique gender values

notable["gender"].value_counts()

notable_gender_other = notable.loc[notable["gender"] == "Other"]

notable_gender_other["birth"].min(), notable_gender_other["birth"].max()

notable_gender_other.loc[notable_gender_other["birth"] == 85.0]

notable.loc[notable["birth"] == .0]

notable["birth"].unique()

notable["birth"] = notable["birth"].astype("Int64")

notable["death"] = notable["death"].astype("Int64")

#notable.info()

# princesses get "prince" as occupation

notable.loc[(notable["level3_main_occ"] == "prince") & 
            (notable["gender"] == "Female")
            ].head()

> Next Steps: 
- look into which columns can be dropped further, e.g. bigperiod_birth/_death columns, areas_of_rattach, pantheon, citizenship, level3_all_occ, ...
- save (then smaller) file
- clean up name column

# check shape of table after dropping columns
notable.shape #reduced no. of columns from 48 to 26

# write new file with reduced no. of columns

# notable.to_csv("../data/notable_people_cross-verified/wiki_notable_people_reduced.csv")
