==> notebooks/data_wrangling/02_wiki_notable_people_wrangling_2.ipynb <==
# Analyzing Wikipedia - Gender-Bias?

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px

# change display settings to show all dataframe columns as well as complete cell values (no truncating values)
# https://stackoverflow.com/questions/47022070/display-all-dataframe-columns-in-a-jupyter-python-notebook
pd.set_option('display.max_columns', None)

# https://stackoverflow.com/a/51540918
# pd.set_option('display.max_rows', 500)
# pd.set_option('display.width', 1000)
pd.set_option('display.expand_frame_repr', False)
pd.set_option('max_colwidth', None)

## Data Source

- **Project**: [A cross-verified database of notable people, 3500BC-2018AD](https://www.nature.com/articles/s41597-022-01369-4)
    - Paper: [Shareable Link](https://rdcu.be/c6YvW)
- **Download**: [A Brief History of Human Time - Cross-verified Dataset](https://data.sciencespo.fr/dataset.xhtml?persistentId=doi:10.21410/7E4/RDAG3O)

Includes data until 2018.

## Read reduced Data

!ls ../data/notable_people_cross-verified/

# read file using utf-8 encoding to accommodate "ä", "ö", ... , "ß" etc.

wiki = pd.read_csv("../data/notable_people_cross-verified/wiki_notable_people_reduced.csv", 
                   encoding='utf-8', index_col=0)

wiki.head()

wiki.info()

# wiki.isnull().sum()

# transform year data from float to integer data type
wiki["birth"] = wiki["birth"].astype("Int64")

wiki["death"] = wiki["death"].astype("Int64")

## First Data Checks based on Notability Index

wiki.sort_values(by="ranking_visib_5criteria").head(10)

wiki.sort_values(by="sum_visib_ln_5criteria", ascending=False).head(10)

# check number of null values

wiki.isnull().sum()

#wiki.info()

## Rename Columns

### Notability Index
Two alternative ways to rank, computed from the same 5 variables below
- ranking_visib_5criteria --> **notability_index_ranking**
- sum_visib_ln_5criteria --> **notability_index_sum**
    - An alternative ranking based on the sum of the log of these variables plus one
    - Slightly different result of ranking

5 variables used to compute notability (with renaming in bold):
- wiki_readers_2015_2018 --> **avg_no_readers_2015_2018**: 
    - the average number of biography views (hits) for each individual between 2015 and 2018 in all available language editions, using an API available [here](https://wikitech.wikimedia.org/wiki/Analytics/AQS/Pageviews) 
    - or zero in the absence of a Wikipedia biography
- **non_missing_score** --> _name stays same_: the number of non-missing items retrieved from Wikipedia or Wikidata for birth date, gender and domain of influence
    - The intuition here is that the more notable the individual, the more documented his/her biographies will be
- total_count_words_b --> **total_count_words**: the length, i.e total number of words found in all available biographies
    - It is equal to zero for individuals with just one Wikidata entry and no biography in Wikipedia
- number_wiki_editions --> **no_wiki_editions**: the number of Wikipedia editions of each individual
- total_noccur_links_b --> **no_external_links**: the total number of external links (sources, references, etc.) from Wikidata

Source: https://www.nature.com/articles/s41597-022-01369-4#Sec2 (section "Measuring notability")

# rename notability index columns
wiki.rename(columns={"sum_visib_ln_5criteria" : "notability_index_sum", 
                     "ranking_visib_5criteria" : "notability_index_ranking",
                     "wiki_readers_2015_2018" : "avg_no_readers_2015_2018",
                     #"non_missing_score" : "",
                     "total_count_words_b" : "total_count_words",
                     "number_wiki_editions" : "no_wiki_editions",
                     "total_noccur_links_b" : "no_external_links"
                    }, inplace=True)

### Occupation Columns

Categories:
- **Discovery/Science**:
    - Academia (Research, Historian, Physician, Scientist, Academic, etc.)
    - Explorer (Engineer, Explorer, Inventor, Sailor, Pioneer, etc.)
- **Culture**:
    - Core (Actor, Writer, Painter, Singer, Music, etc.)
    - Periphery (Journalist, Architect, Model, Designer, Presenter, etc.)
- **Leadership**:
    - Politics (Politician, Activist, Revolutionary, Trade unionist, Minister, etc.)
    - Military (Military, Officer, Commander, Soldier, Army, etc.)
    - Law (Lawyer, Diplomat, Judge, Jurist, Civil service)
    - Nobility (Aristocrat, Noble, King, Sovereign, Monarch, etc.)
    - Religious (Priest, Prelate, Rabbi, Missionary, Bishop, etc.)
    - Corporate Leadership (Business, Entrepreneur, Bank, Merchant, Manager, etc.)
- **Sports/Games**: 
    - Football, Player, Sport, Baseball, Basket, etc.
- **Other**:
    - Worker (Farmer, Librarian, Musher, Bookseller, Printer, etc.)
    - Family (Son, Daughter, Child, Wife of, Father, etc.)
    - Misc. (Esperantist, Criminal, Convict, Killer, Philanthropist, etc.)

Source: https://www.nature.com/articles/s41597-022-01369-4#Sec2 (section "Domains of influence and occupations")

Renaming:
- level1_main_occ --> **occup_l1**
- level2_main_occ --> **occup_l2**
- level3_main_occ --> **occup_l3**
- level3_all_occ  --> **occup_l3_all**

# rename occupation columns
wiki.rename(columns={"level1_main_occ" : "occup_l1", 
                     "level2_main_occ" : "occup_l2",
                     "level3_main_occ" : "occup_l3",
                     "level3_all_occ"  : "occup_l3_all"
                    }, inplace=True)

# wiki.info()

### Other Columns

Geolocalisation: _using longitudes and latitudes of birth and death places_
- bplo1 --> **birth_place_lon**
- dplo1 --> **death_place_lon**
- bpla1 --> **birth_place_lat**
- dpla1 --> **death_place_lat**

Periods of human history:
- bigperiod_birth_graph_b --> **bigperiod_birth**
- bigperiod_death_graph_b --> **bigperiod_death**

Citizenship:
- citizenship_1_b --> **citizenship**

Source: https://www.nature.com/articles/s41597-022-01369-4#Sec2 

# rename other columns
wiki.rename(columns={"bplo1" : "birth_place_lon", 
                     "dplo1" : "death_place_lon",
                     "bpla1" : "birth_place_lat",
                     "dpla1" : "death_place_lat",
                     "bigperiod_birth_graph_b" : "bigperiod_birth",
                     "bigperiod_death_graph_b" : "bigperiod_death",
                     "citizenship_1_b" : "citizenship"
                    }, inplace=True)

wiki.info()

#wiki.head(10)

## Clean "name" Column

# replace _underscores with whitespace
wiki["name"] = wiki["name"].str.replace("_", " ")

# remove everything in () behind actual name
wiki["name"] = wiki["name"].str.replace(r" \(.*\)", "")

## Clean "citizenship" Column

wiki["citizenship"].unique()

# replace _underscores with whitespace
wiki["citizenship"] = wiki["citizenship"].str.replace("_", " ")

#wiki.sample(50)

## Check unique Values and Counts after Data Cleaning

wiki["citizenship"].unique()

wiki["birth"].unique()

wiki["death"].unique()

wiki["gender"].unique()

wiki["occup_l1"].unique()

wiki["occup_l1"].value_counts()

wiki["occup_l2"].unique()

wiki["occup_l2"].value_counts()

wiki["occup_l3"].value_counts()

wiki["un_subregion"].unique()

wiki["un_region"].unique()

wiki["bigperiod_birth"].value_counts()

wiki["bigperiod_death"].value_counts()

wiki["non_missing_score"].unique()

wiki["notability_index_ranking"].unique()

## Write cleaned Data into new File

wiki.shape

#wiki.to_csv("../data/notable_people_cross-verified/wiki_notable_people_cleaned.csv")
