==> notebooks/data_wrangling/03_wiki_notable_people_iso_codes_1.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px

# change display settings to show all dataframe columns as well as complete cell values (no truncating values)
# https://stackoverflow.com/questions/47022070/display-all-dataframe-columns-in-a-jupyter-python-notebook
pd.set_option('display.max_columns', None)

# https://stackoverflow.com/a/51540918
# pd.set_option('display.max_rows', 500)
# pd.set_option('display.width', 1000)
pd.set_option('display.expand_frame_repr', False)
pd.set_option('max_colwidth', None)

## Read Data

# read data from cleaned file
!ls ../data/notable_people_cross-verified/

# create dataframe from cleaned, reduced file

wiki = pd.read_csv("../data/notable_people_cross-verified/wiki_notable_people_cleaned.csv")

wiki.head()

## Add ISO codes for Countries of Citizenship using Country Converter Package

> [Link](https://pypi.org/project/country-converter/) to package

# import country converter package

import country_converter as coco

coco.convert(names=["Germany"], to="ISO3")

# create empty list for new dataframe series

citizenship_alpha_3 = []

wiki["citizenship"]

c = ["Czech Republic", "Germany"]

print(coco.convert(names=c, to="ISO3"))

# testing functionality of package with smaller sample

# iso3 = [coco.convert(names=wiki["citizenship"].sample(50), to="ISO3")]

#iso3

### Creating new Column "citizenship_iso3" and adding ISO3 Code per Row

> Very slow. Next time develop a more performant solution for this dataset.

# wiki["citizenship_iso3"] = coco.convert(names=wiki["citizenship"], to="ISO3", not_found=None)

wiki.head()

# Write new file including ISO3 codes

# wiki.to_csv("../data/notable_people_cross-verified/wiki_notable_people_cleaned_iso.csv")
